+++ runtime-api/bonded-coins/src/pool_details.rs
